Give more memory to processes
A couple processes were running out of memory so I've given them more. I
also got an error around singularity not being found, which is odd given
it should be loaded by the scw profile... I've added a load for that and
lastly added a new nextflow run line for trying just our data (not the
yang 2022 data).

I added that in case there's any issues that arise due to the yang data,
but it's running well so far, currently on the integrate step.