Update to parallel
It seems this process takes a very long time, over 30 mins per file. Since there's a file for every geneset per chromosome, there's a lot of files.
I've also changed this to only use the top 10% of genes to reduce the number of files.
Hoping the change to highmem which has more nodes will also help